flask_app/requirements.txt

blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==3.0.3
SQLAlchemy==1.4.23
pymysql==1.0.2
sshtunnel==0.4.0
mysql-connector-python
Flask-SQLAlchemy
